:hammer: chore: add `@snailycad/utils` package
