+++ packages/api/src/controllers/dispatch/StatusController.ts
@@ -29,7 +29,7 @@ import {
   Value,
   WhitelistStatus,
 } from "@prisma/client";
-import { generateCallsign } from "utils/callsign";
+import { generateCallsign } from "@snailycad/utils";
 import { validateSchema } from "lib/validateSchema";
 import { handleStartEndOfficerLog } from "lib/leo/handleStartEndOfficerLog";
 
@@ -218,7 +218,6 @@ export class StatusController {
   async mergeOfficers(
     @BodyParams("id") id: string,
     @Context("activeOfficer") activeOfficer: Officer,
-    @Context("cad") cad: cad & { miscCadSettings: MiscCadSettings },
   ) {
     if (id === activeOfficer.id) {
       throw new BadRequest("officerAlreadyMerged");
@@ -252,13 +251,10 @@ export class StatusController {
       select: { id: true },
     });
 
-    // todo: fix type
-    const callsign = generateCallsign(activeOfficer as any, cad.miscCadSettings.pairedUnitTemplate);
-
     const unit = await prisma.combinedLeoUnit.create({
       data: {
         statusId: status?.id ?? null,
-        callsign,
+        callsign: activeOfficer.callsign,
       },
     });
 
@@ -339,9 +335,8 @@ function createWebhookData(webhook: APIWebhook, miscCadSettings: MiscCadSettings
 
   const status = unit.status?.value.value ?? "Off-duty";
   const unitName = isNotCombined ? `${unit.citizen.name} ${unit.citizen.surname}` : "";
-  const callsign = isNotCombined
-    ? generateCallsign(unit, miscCadSettings.callsignTemplate)
-    : unit.callsign;
+  // todo: fix type
+  const callsign = generateCallsign(unit as any, miscCadSettings.callsignTemplate);
   const officerName = isNotCombined
     ? `${unit.badgeNumber} - ${callsign} ${unitName}`
     : `${callsign}`;
