+++ packages/client/package.json
@@ -30,6 +30,7 @@
     "@snailycad/config": "1.0.0-beta.59",
     "@snailycad/schemas": "1.0.0-beta.59",
     "@snailycad/types": "1.0.0-beta.59",
+    "@snailycad/utils": "1.0.0-beta.59",
     "axios": "^0.26.0",
     "date-fns": "^2.28.0",
     "formik": "^2.2.9",
