+++ packages/client/src/hooks/useGenerateCallsign.ts
@@ -1,53 +1,15 @@
 import { useAuth } from "context/AuthContext";
-import type { FullDeputy, FullOfficer } from "state/dispatchState";
-import type { CombinedLeoUnit } from "@snailycad/types";
+import type { CombinedLeoUnit, EmsFdDeputy, MiscCadSettings, Officer } from "@snailycad/types";
+import { generateCallsign } from "@snailycad/utils";
 
-type P = "callsign" | "callsign2" | "department";
-type FullUnit =
-  | Pick<FullOfficer, P | "divisions">
-  | Pick<FullDeputy, P | "division">
-  | CombinedLeoUnit;
+type P = "callsign" | "callsign2" | "department" | "citizenId";
+type Unit = Pick<Officer, P | "divisions"> | Pick<EmsFdDeputy, P | "division"> | CombinedLeoUnit;
+type TemplateId = keyof Pick<MiscCadSettings, "pairedUnitTemplate" | "callsignTemplate">;
 
 export function useGenerateCallsign() {
   const { cad } = useAuth();
   const miscCadSettings = cad?.miscCadSettings;
 
-  function generateCallsign(unit: FullUnit) {
-    if (!("department" in unit)) {
-      return "NULL";
-    }
-
-    const { callsign, callsign2, department } = unit;
-    const unitDivision =
-      ("division" in unit && unit.division) || ("divisions" in unit ? unit.divisions : []);
-
-    const template = miscCadSettings?.callsignTemplate;
-    const [division] = Array.isArray(unitDivision) ? unitDivision : [unitDivision];
-
-    if (!template) {
-      return "";
-    }
-
-    const replacers = {
-      department: department?.callsign,
-      callsign1: callsign,
-      callsign2,
-      division: division?.callsign,
-    };
-
-    const templateArr: (string | null)[] = template.split(/[{}]/);
-    Object.entries(replacers).forEach(([replacer, value]) => {
-      const idx = templateArr.indexOf(replacer);
-
-      if (value) {
-        templateArr[idx] = value;
-      } else {
-        templateArr[idx] = null;
-      }
-    });
-
-    return templateArr.filter((v) => v !== null).join("");
-  }
-
-  return generateCallsign;
+  return (unit: Unit, templateId: TemplateId = "callsignTemplate") =>
+    generateCallsign(unit, miscCadSettings?.[templateId] ?? null);
 }
