+++ packages/utils/README.md
@@ -1,0 +1,3 @@
+# @snailycad/utils
+
+TODO
